Notes endpoint integration

+++ src/Views/Common/Docs/Notes.jsx
@@ -7,7 +7,7 @@ import ButtonComponent from "Components/Button/Button";
 import SpinnerComponent from "Components/Spinner/Spinner";
 import { OverallModel } from "Views/Students/Utils/OverallModal";
 import { useCommonState, useDispatch } from "Components/CustomHooks";
-import {setEditNoteData, updateModalShow} from "Views/Common/Slices/Common_slice"
+import {setEditNoteData, showMoreModal, updateModalShow} from "Views/Common/Slices/Common_slice"
 import { getStudentsNotes, deleteStudentNotes ,updateNotePriority } from "Views/Common/Actions/Common_action";
 import 'Stylesheet/Css/StudentsNotes.css'
 
@@ -23,22 +23,30 @@ const Notes = () => {
   }, []);
 
    const handleDeleteNote = (id) => {
-    dispatch(deleteStudentNotes(id));
-  };
+    if (window.confirm("Are you sure you want to delete this note?")) {
+        dispatch(deleteStudentNotes(id));
+      };
+    }
+    
 
   const handlepriority = (id) =>{
     dispatch(updateNotePriority(id))
   }
 
   const handleEdit = (note_data) =>{
-      dispatch(setEditNoteData({id: note_data.id, title: note_data.title, content: note_data.content}));
+      dispatch(setEditNoteData({id: note_data.id, title: note_data.title, content: note_data.notes}));
       dispatch(updateModalShow({ show: true, close_btn: true, modal_from: "Notes", modal_type: "edit_notes",}));
   }
 
- 
+  const handleShowMore = (note_data) => {
+  dispatch(showMoreModal({id: note_data.id, title: note_data.title,content: note_data.notes}));
+  dispatch(updateModalShow({show: true, close_btn: true,modal_from: "Notes",modal_type: "show_more_note"}));
+}
+
+
 
   return (<>
-    <div className="row  px-3">
+    <div className="container-fluid">
       <div className="w-100  border-bottom pb-3">
         <div className="col d-flex align-items-center justify-content-between">
           <h4 className="m-0">My Notes</h4>
@@ -54,16 +62,16 @@ const Notes = () => {
           <div className="w-100  row align-items-center justify-content-center mt-5">
             <div className="col-6 text-center">
               <SpinnerComponent />
-              <p className="py-3">Geeting Notes</p>
+              <p className="py-3">Getting Notes</p>
             </div>
           </div>
         ) : data?.length ? (
           <div className="w-100 small_header_content_main row overflowY">
          { data?.map((note_data) => (
-               <div className="col-md-6 col-lg-4 col-xxl-3 p-2 " key={note_data.id}>
+               <div className="col-md-6 col-lg-4 col-xxl-3 p-2" key={note_data.s_no}>
                  <NoteCard 
                  data={note_data} 
-                 onShowMore={(message) => {console.log("notes", message)}} 
+                 onShowMore={() => handleShowMore(note_data)} 
                  notesDeleteOnClick={() => handleDeleteNote(note_data.id)}
                  addFavoriteOnClick={()=> handlepriority(note_data.id)}
                  noteFavoriteIcon={note_data.priority === "high" ? Icons.NoteStarIcon : Icons.favorite_outline_icon}
